requirements.txt
python3-nmap 
urllib
validators
termcolor
socket
